refactor: move api related lib and models to new server folder
